Preparations for the implementation of the plugin framework. This has moved some methods to other parts of the architecture. But these placements are perhaps transitory; let's discuss them. The ends of the architecture (PTI and Datadock) has been retrofitted to use the pluginframework. But not completely and not in a very generic manner (In fact just the opposite...).
Also, a lot of tests has been edited and ignored. These are severe changes and must be amended as soon as the dust has settled.

CargoObject now has the ability of returning a byte[] of the underlying List<Byte>

FedoraHandler should be disposed at first opportunity. It contains just comments.


git-svn-id: svn://oss.dbc.dk/sync_search/opensearch/trunk@3265 51cac136-f83d-432d-bb6f-125e83a60a2d