Added new target to build file: dist-aux which builds all the auxillary targets (helper applications)
renamed the all option in dist.py back to both, as there are no need to distribute auxillary applications.

Added the IndexChecker package which contains a classes for checking indexable files against a predefined index (it indexes the indexable files and check them against the predefined result.)

Results and tests runs are facilitated through the IndexerMain, while unittest integration are facilitated  through the IndexCheckerTest.

Added an example test folder




git-svn-id: svn://oss.dbc.dk/sync_search/opensearch/trunk@4329 51cac136-f83d-432d-bb6f-125e83a60a2d