Made the following changes:
Made new ant target: dist-aux, which builds distros of the auxillary
 applications (indexchekcer, readindex, testindexer ) The dist target
 only creates the pti and datadock now.
 
Applied changes to dist_py script, so it is only the datadock and pti
which can be distributed and hence, noth replaces all as dist target

Applied the IEstimate interface to the Indexers in the ptiThread. The
IndexerXSEM now uses the IEstimate handed down through the ptithread
instead of making a new instance (which is very naughty)

Made the indexchecker package which is made accesible through a main
method, and a unittest class. With the indexchecker jar file it is
possible to check some indexable files to a indexresult, and see of
they match, the application can also be used to generate result
files. The Test class is called during unittests and tests the harvest
folder found in trunk/tests/testdata. 

The documentation is lacking, which I will fix after this commit






git-svn-id: svn://oss.dbc.dk/sync_search/opensearch/trunk@4335 51cac136-f83d-432d-bb6f-125e83a60a2d