Rolled back the static calls in fedoraAdministration due to neccessesity of dependency injection for the indexTester and indexTest executables, and added the new non-static methods to the IFedoraAdministration interface.
The fedoraAdministration instance is made in the main methods, and are propegated to the dependent classes through their constructors

The FedoraCommunication class has been refactored into the FedoraAdministration and I have changed the names of the unittests to reflect this.

The Indexer files have been moved from src to functionaltests foldertree, which had the effect that the needed class files where not compiled at the time of jar assembly. The build file now compiles the functionaltests tree classes before trying to assembly dist-aux targets 



git-svn-id: svn://oss.dbc.dk/sync_search/opensearch/trunk@4710 51cac136-f83d-432d-bb6f-125e83a60a2d