Made the src ready for oracle communication for the new Harvester.
Firstly a new interface ( IDBConnection ) is added and the old DBConnection is now called postgresqlDBConnection and implements the mentioned interface. A new oracle implementation is added, and the relevant class is instanced in the main methods and propagated through to the user classes through DP. The configuration framework is adjusted, so it now contains db configuration for both postgresql and oracle.

The unittests for Processqueue and estimate reflects the change as well



git-svn-id: svn://oss.dbc.dk/sync_search/opensearch/trunk@4717 51cac136-f83d-432d-bb6f-125e83a60a2d